一旦無理やり条件つけてバグらないようにして、マージしたいｗ
